added common base class for input bridge
